Removed sample dir and cleaned up

+++ packages/qubo_ahrs/src/config.cpp
@@ -11,7 +11,7 @@ int main(int argc, char *argv[])
          IMU *imu = new IMU(std::string(argv[1]),getBaudrate(argv[2]));
          IMU::AcqConfig config;
          IMU::FilterData filters;
-         unsigned int i;
+         long unsigned int i;
 
          imu->openDevice();
          printf("Connected to %s.\n", (imu->getInfo()).c_str());
@@ -26,9 +26,9 @@ int main(int argc, char *argv[])
          printf("Sample delay:      %f\n", config.sample_delay);
 
          filters = imu->getFIRFilters();
-         printf("Number of filters: %u\n", filters.size());
+         printf("Number of filters: %lu\n", filters.size());
          for (i=0; i < filters.size(); i++)
-         printf("Filter #%.2d:        %f\n",i, filters[i]);
+         printf("Filter #%.2lu:        %f\n",i, filters[i]);
 
          printf("Mag Truth Method:  %d\n", imu->getMagTruthMethod());
          printf("Functional mode:   %d\n", imu->getAHRSMode());
